Incorporated the review comments

+++ CrackingCodingInterview/vAnil/05 Bit Manipulation/03 FlipBitToWin/FlipBitToWin.cpp
@@ -18,7 +18,7 @@ current = 0
 
 */
 
-MY03FLIPBITTOWIN_API int Longest1Sequence(unsigned num)
+MY03FLIPBITTOWIN_API int Longest1Sequence(size_t num)
 {
 	int max = 0;
